Remove Authorization dependency from Application project
