+++ Application/Queries/CharacterQueries/GetCharacterOwnerEmailQuery.cs
@@ -1,0 +1,2 @@
+﻿namespace Application.Queries.CharacterQueries;
+public record GetCharacterOwnerEmailQuery(int CharacterId) : IRequest<string?>;
