+++ Application/Queries/CountryQueries/GetCountryOwnerEmailQuery.cs
@@ -1,0 +1,2 @@
+﻿namespace Application.Queries.CountryQueries;
+public record GetCountryOwnerEmailQuery(int CountryId) : IRequest<string?>;
